90% on deletion and got rid of duplicate names
Made it so that duplicate names are not allowed for groups/cams after we fixed the bug on deletion that wouldn't update the list. BUT for some reason after this addition, it no longer works. Need to investigate.
Also added deleteCamera call when the camera registration fails so that it is deleted in the backend.