Added the delete button for Groups and Cameras Screens plus the duplicate name check (#75)
* Partial Completion of AppSide work

Includes changes to the Setting Wheel to Trashcan icon
Added some of the alert popups. Need to add the one for if a group still has cameras in it.

* Delete button backend logic

Updated the service files in the backend for the delete groups and cameras
Added some `steps.ts` test files for the functionality. Need to check them on desktop and then begin working on the controller and connecting it to the App.

* Initial functionality/testing/app-work partway done

Needs to finish testing on `delete-group.steps.ts`
Then ensure that the buttons on the app work so that the camera/notifs and group are deleted when conditions are met

* Refactored the DeleteButton to two separate files

The delete button now has a button for just groups and just cameras. looks the same but splits up the functionality for better clarity

* Fixed a hook error and some small styling changes

Updated the screens so that they have the correct buttons applied.
There is an issue where the delete button isn't doing anything at all. Need to address that issue as well as fix the backend testing and making sure the backend APIs work.

* Added delete notification backend func and edits to testing

Added some changes to how the app processes what the backend returns when deleting groups/cameras. Tried to fix the backend tests but to no avail.

* Recommiting for backend

* Pushing for tucker

helping tucker

* push for backend

testing to see how delete Camera is getting along

* push again for camera removal

got rid of an unncessary call to deleteNotificationAPI

* Updating delete URL for deleting camera

* Trying to get rid of delete cam 404 and cleaning up

* Adding more params to setup for later work. and to reupload backend.

* log to show what is being sent.

* checking to see if removeCamera is being called

* 90% on deletion and got rid of duplicate names

Made it so that duplicate names are not allowed for groups/cams after we fixed the bug on deletion that wouldn't update the list. BUT for some reason after this addition, it no longer works. Need to investigate.
Also added deleteCamera call when the camera registration fails so that it is deleted in the backend.

* Attempt to fix tests

* Tests fixed! (almost)

* Final touchups

-Fixed some app-lint errors
-Fixed logic errors and added some more error logging / alerts on the app side.
-Added a new test for when someone tries to delete someone else's camera and made the tests match the steps and feature files they came from.

* Final changes to make app-lint pass

should fix the issues

* Applying comments from PR

* Fixing change to backend test

---------

Co-authored-by: Avery Black <avery.c.black@gmail.com>